#5 refactoring
